hyperthreading: ensure hf_ncores as the number of cores in a VSI
Ajith reported that we need to set hf_ncores independent of
hyperthreading_enabled. This commit fixes the misconfiguration at
vpc.tf

Unlike LSF, we do not set EGO_DEFINE_NCPUS since users can be aware
of it and that config is often not used.

From Ajith:

EGO_DEFINE_NCPUS=threads is not the default setting for Symphony and
most customers don't use it. The reason a customer may define "threads"
is to handle a heterogeneous environment with CPU's that have HT enabled
and others that don't have HT available. If the environment is
homogeneous and all machines have HT enabled, then there is no need to
enable "thread" mode.

We can enable HT in the hardware but we should still define a "CPU" as a
core as this is what most customers will do.

> Symphony will create two host factory nodes for both
> hyperthreading_enabled=true and =false.
> Am I correct?

Yes, host factory by default provisions cores and not threads, so the
$hf_ncores value should always be the number of cores in a VSI and not
threads.

Symphony provisions slots which is a minimum resource unit for
scheduling. A slot contains some number of cores and memory. If the
customer wants to account for HT, then they can define a slot as 1/2 a
core and then workload can be scheduled to one thread. So the core
setting in Host Factory is not something that needs to account for HT.

Reference: https://github.ibm.com/workload-eng-services/HPCCluster/issues/571
Reported-by: Ajith Shanmuganathan <ashanmug@ca.ibm.com>