Prod Bug: Error: No image found with name hpcc-lsf10-cent77-jul2221-v6
This code change is to fix issue#694. workload-eng-services/HPCCluster#694

@TYOS As I explained, the problem can only be experience with non-dev
ibm cloud account where the data block lookup of
data.ibm_is_image.(var.image_name) will fail as the image name is not
visible to non-dev account. Problem is even if we have a entry in the
imagename_image_id mapping fail, this code to lookup the image with
var.image_name is being executed and causing the failure irrespective of
the entry there in the mapping file or not.